refactor(e2e): split the mcp gateway suite into feature-scoped specs with a parametrized auth-header matrix
